+++ src/util/mod.rs
@@ -1,2 +1,3 @@
 pub mod alpha_bleed;
+pub mod animation;
 pub mod svg;
